+++ packages/freezed/example/lib/main.freezed.dart
@@ -1,5 +1,6 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'main.dart';
 
@@ -175,7 +176,8 @@ Union _$UnionFromJson(Map<String, dynamic> json) {
       return Complex.fromJson(json);
 
     default:
-      throw FallThroughError();
+      throw CheckedFromJsonException(json, 'custom-key', 'Union',
+          'Invalid union type "${json['custom-key']}"!');
   }
 }
 
@@ -302,7 +304,7 @@ class _$DataCopyWithImpl<$Res> extends _$UnionCopyWithImpl<$Res>
 class _$Data with DiagnosticableTreeMixin implements Data {
   const _$Data(this.value);
 
-  factory _$Data.fromJson(Map<String, dynamic> json) => _$_$DataFromJson(json);
+  factory _$Data.fromJson(Map<String, dynamic> json) => _$$DataFromJson(json);
 
   @override
   final int value;
@@ -391,7 +393,7 @@ class _$Data with DiagnosticableTreeMixin implements Data {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$DataToJson(this)..['custom-key'] = 'Default';
+    return _$$DataToJson(this)..['custom-key'] = 'Default';
   }
 }
 
@@ -427,7 +429,7 @@ class _$Loading with DiagnosticableTreeMixin implements Loading {
   const _$Loading();
 
   factory _$Loading.fromJson(Map<String, dynamic> json) =>
-      _$_$LoadingFromJson(json);
+      _$$LoadingFromJson(json);
 
   @override
   String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
@@ -502,7 +504,7 @@ class _$Loading with DiagnosticableTreeMixin implements Loading {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$LoadingToJson(this)..['custom-key'] = 'Loading';
+    return _$$LoadingToJson(this)..['custom-key'] = 'Loading';
   }
 }
 
@@ -549,7 +551,7 @@ class _$ErrorDetails with DiagnosticableTreeMixin implements ErrorDetails {
   const _$ErrorDetails([this.message]);
 
   factory _$ErrorDetails.fromJson(Map<String, dynamic> json) =>
-      _$_$ErrorDetailsFromJson(json);
+      _$$ErrorDetailsFromJson(json);
 
   @override
   final String? message;
@@ -638,7 +640,7 @@ class _$ErrorDetails with DiagnosticableTreeMixin implements ErrorDetails {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$ErrorDetailsToJson(this)..['custom-key'] = 'Error';
+    return _$$ErrorDetailsToJson(this)..['custom-key'] = 'Error';
   }
 }
 
@@ -694,7 +696,7 @@ class _$Complex with DiagnosticableTreeMixin implements Complex {
   const _$Complex(this.a, this.b);
 
   factory _$Complex.fromJson(Map<String, dynamic> json) =>
-      _$_$ComplexFromJson(json);
+      _$$ComplexFromJson(json);
 
   @override
   final int a;
@@ -790,7 +792,7 @@ class _$Complex with DiagnosticableTreeMixin implements Complex {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$ComplexToJson(this)..['custom-key'] = 'Complex';
+    return _$$ComplexToJson(this)..['custom-key'] = 'Complex';
   }
 }